refactor: removed print from eval functions
